requirements.txt
langchain-openai>=0.0.5
python-dotenv>=1.0.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1 